+++ Sources/ClientRuntime/Retries/DefaultRetryErrorInfoProvider.swift
@@ -6,6 +6,10 @@
 //
 
 import struct Foundation.TimeInterval
+import struct SmithyRetriesAPI.RetryErrorInfo
+import enum SmithyRetriesAPI.RetryErrorType
+import protocol SmithyRetriesAPI.RetryErrorInfoProvider
+
 
 public enum DefaultRetryErrorInfoProvider: RetryErrorInfoProvider {
