+++ Tests/SmithyRetriesTests/RetryIntegrationTests.swift
@@ -7,6 +7,8 @@
 
 import Foundation
 import XCTest
+import SmithyRetriesAPI
+@testable import SmithyRetries
 @testable import ClientRuntime
 
 // This test class reproduces the "Standard Mode" test cases defined in "Retry Behavior 2.0"
